Move viewport event into seperate file
